requirements.txt
fastapi
uvicorn[standard]
motor
python-dotenv
pydantic[email]
cloudinary
razorpay
sendgrid
PyJWT
bcrypt
python-multipart

